+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Removed redundant code that gave error when closing the database
 
 ## [0.0.5] - 2024-01-12
 ### Fixed
@@ -31,4 +33,3 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [0.0.4]: https://github.com/adnaniazi/fast5_rekindler/compare/0.0.3...0.0.4
 [0.0.3]: https://github.com/adnaniazi/fast5_rekindler/compare/0.0.2...0.0.3
 [0.0.2]: https://github.com/adnaniazi/fast5_rekindler/tree/0.0.2
-